Moved reused rugged strToInt method to Util.  Fix in PdfPrinter to handle front chapter id.
